api/requirements.txt
fastapi
python-dotenv
qrcode
azure-storage-blob
